+++ lib/lydown/parsing.rb
@@ -23,7 +23,6 @@ class LydownParser
 
     unless ast
       error_msg = format_parser_error(source, parser, opts)
-      $stderr.puts error_msg
       raise LydownError, error_msg
     else
       stream = []
@@ -54,7 +53,7 @@ class LydownParser
   end
 
   def self.format_parser_error(source, parser, opts)
-    msg = opts[:filename] ? "#{opts[:filename]}: " : ""
+    msg = opts[:filename] ? "#{Pathname.relative_pwd(opts[:filename])}: " : ""
     if opts[:nice_error]
       msg << "Unexpected character at line #{parser.failure_line} column #{parser.failure_column}:\n"
     else